We have a basic functional implementation of a neural network training process with two hidden layers.
What essentially is happening is using open cv, we convert an image to a 3d matrix and to resize it into desired row and column size, using the resized 3d matrix we transform it into a 2d matrix with get_sum_pixels method. The resulting matrix is then converted to type double and fed into neural network forward and backward methods as 'x'.

Still need to test and verify all parts involved, as well as work on a better result ,a.k.a, y matrix generation.