Finalized initial version of fit method, where actual training of neural network will occur.
For now network architecture is dynamic, where user can create how many layers necessary based on the classes implemented.
We store class sequence on a vector in neural network class and iterate through them in forward_pass method.

Fit method returns 2d matrix which is default output typing for output layer, receiving as parameters training data and epochs, number of times forward and backward propagation will be performed.

Weights and biases for each layer are being created based on flags in base class construction, to determine wether or not layer has weights and biases, if so they are added as std::any data types and stored in vectors inside neural network class.
For now weights and biases for convolutional and dense layers are functional and passing correct weights and biases throughout epochs.
Next commit will be focused on backward propagation to alter weights and biases created in this commit so the network can acctually learn and create ideal set of weigths and biases to be added exported model.

Also implemented sigmoid activation function.