We are finally seing some signs of learning for our network.
Wrote a ton of methods in other classes, to support backward propagation.

For now, it's a little hard coded to our example network architecture but with only dense and flattening layer backward propagation implemented my network is already showing signs of learning if the epoch is high enough.

Will continue with backward propagation for next layers on upcoming commits.