Let go a little of multithreading in order to get export_model going.
We have created export_model and load_model which are currently working with high stability, although there are a lot of hard coded parts.

But we have achieved our goal, we are currently able to write all relevant data regarding the network in a binary file, and with load model we completely recreated stored data.